// src/api/mobile/models/ReportFotosRes.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export enum ReportFotosResReportTypeEnum {
  QUALITY = 'QUALITY',
  IMAGE_PACK = 'IMAGE_PACK',
}

export type ReportFotosRes = {
  /**
   * @type string
   */
  _id: string;
  /**
   * @type string
   */
  reportType: ReportFotosResReportTypeEnum;
  /**
   * @type string
   */
  appName: string;
  /**
   * @type string | undefined
   */
  appVersion?: string;
  /**
   * @type string | undefined
   */
  userId?: string;
  /**
   * @type string | undefined
   */
  correlationId?: string;
  /**
   * @type string
   */
  status: string;
  /**
   * @type number
   */
  attempts: number;
  /**
   * @type string | undefined
   */
  technicalError?: string;
  /**
   * @type array
   */
  media: object[];
  /**
   * @type object
   */
  properties: {
    [key: string]: any;
  };
  /**
   * @type string, date-time
   */
  createdAt: Date;
  /**
   * @type string, date-time
   */
  updatedAt: Date;
  /**
   * @type string | undefined, date-time
   */
  uploadedAt?: Date;
};